build: add arch defs and enhance build tool detection
Add architecture preprocessor definitions (_WIN64, _AMD64_, _X86_) in CMakeLists.txt for MSVC builds, which are required for kernel driver compilation. Update virtual_mic.h to change BufferUsage from FLOAT to ULONG (percentage) and remove the unused wdf.h include. Enhance build.ps1 by translating all comments and output to English, and implementing fallback logic to locate CMake and MSBuild in standard Visual Studio installation paths if they are not found in the system PATH, improving build environment robustness.

+++ include/virtual_mic.h
@@ -2,7 +2,6 @@
 #define VIRTUAL_MIC_H
 
 #include <ntddk.h>
-#include <wdf.h>
 #include <ntstrsafe.h>
 
 // IOCTL Codes
@@ -36,7 +35,7 @@ typedef struct _SET_FORMAT_REQUEST {
 typedef struct _DRIVER_STATS {
     BOOLEAN IsActive;
     ULONG64 SamplesProcessed;
-    FLOAT BufferUsage;
+    ULONG BufferUsage; // Percentage as integer (0-100)
     ULONG Underruns;
     ULONG Overruns;
     AUDIO_FORMAT CurrentFormat;
